app/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
asyncpg>=0.29.0
databricks-sdk>=0.20.0
databricks-sql-connector>=3.0.0
httpx>=0.25.0
python-multipart>=0.0.6
